pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pr-sop"
version = "0.1.0"
description = "Opinionated PR governance checks. Config-driven, no LLM. Runs as pre-commit hook, CLI, or GitHub Action."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Pawan Singh Kapkoti" }]
keywords = ["pr", "governance", "pre-commit", "github-action", "devtools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
pr-sop = "pr_sop.cli:app"

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.hatch.build.targets.wheel]
packages = ["pr_sop"]

[tool.ruff]
line-length = 100
